pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "miband2auth"
authors = [{name = "Anton Ptashnik", email = "iavtomator@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.7"
dependencies = [
    "pyaes >=1.6",
    "dbus-python >= '1.2'",
]

[project.scripts]
miband2auth = "miband2auth.main:main"
